+++ ChANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Commit 08-07-2022 (v2.5.2)
+
+- Add `resource()` method to `Horizom\Routing\RouteCollector`.
+- Add `resources()` method to `Horizom\Routing\RouteCollector`.
+- Add `redirect()` method to `Horizom\Routing\RouteCollector`.
+- Add `redirectPermanently()` method to `Horizom\Routing\RouteCollector`.
+
 ## Commit 08-03-2022 (v2.5.0)
 
 - Implement request file upload.